pyproject.toml
[project]
name = "skills-as-mcp"
version = "0.1.0"
description = "MCP server that provides Claude Skills-like functionality via Model Context Protocol"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp[cli]>=1.2.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "python-frontmatter>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.1",
]

[project.scripts]
skills-as-mcp = "skills_as_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/skills_as_mcp"]

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
